Refactor render.Targets
